+++ DATA KERJA/PROYEK BUANA MEKAR/buanaMekar/src/main/java/com/example/buanaMekar/controllers/produkController.java
@@ -56,7 +56,7 @@ public class produkController {
     }
     
     @RequestMapping("/produk/edit/{id}")
-    public ModelAndView showEditProdukForm(@PathVariable(name = "id")Long id, Model model){
+    public ModelAndView showEditProdukForm(@PathVariable(name = "id")Integer id, Model model){
         ModelAndView mav = new ModelAndView("editProduk");
         Produk produk = service.get(id);
         model.addAttribute("jenisProduk", service.getAllJenisProduk());
@@ -65,7 +65,7 @@ public class produkController {
     }
     
     @RequestMapping("/produk/delete/{id}")
-    public String deleteProduk(@PathVariable(name = "id")Long id){
+    public String deleteProduk(@PathVariable(name = "id")Integer id){
         service.delete(id);
         
         return "redirect:/produk";
