+++ DATA KERJA/PROYEK BUANA MEKAR/buanaMekar/src/main/java/com/example/buanaMekar/services/ProdukService.java
@@ -38,12 +38,12 @@ public class ProdukService {
 
     }
 
-    public Produk get(long id) {
+    public Produk get(Integer id) {
         return repo.findById(id).get();
 
     }
 
-    public void delete(long id) {
+    public void delete(Integer id) {
         repo.deleteById(id);
 
     }
